requirements.txt
gradio
requests
pillow
matplotlib
pandas
numpy
python-dotenv
selenium
tabulate
